working on ErrorHandler
