+++ src/Viserio/Exception/Handler.php
@@ -2,8 +2,6 @@
 declare(strict_types=1);
 namespace Viserio\Exception;
 
-use Error;
-use ErrorException;
 use Interop\Container\ContainerInterface;
 use Interop\Http\Factory\ResponseFactoryInterface;
 use Narrowspark\HttpStatus\Exception\AbstractClientErrorException;
@@ -19,28 +17,20 @@ use Symfony\Component\Debug\DebugClassLoader;
 use Symfony\Component\Debug\Exception\FatalErrorException;
 use Symfony\Component\Debug\Exception\FatalThrowableError;
 use Symfony\Component\Debug\Exception\FlattenException;
-use Symfony\Component\Debug\Exception\OutOfMemoryException;
-use Symfony\Component\Debug\FatalErrorHandler\ClassNotFoundFatalErrorHandler;
-use Symfony\Component\Debug\FatalErrorHandler\UndefinedFunctionFatalErrorHandler;
-use Symfony\Component\Debug\FatalErrorHandler\UndefinedMethodFatalErrorHandler;
 use Throwable;
 use Viserio\Contracts\Config\Repository as RepositoryContract;
 use Viserio\Contracts\Config\Traits\ConfigAwareTrait;
-use Viserio\Contracts\Container\Traits\ContainerAwareTrait;
 use Viserio\Contracts\Exception\Displayer as DisplayerContract;
 use Viserio\Contracts\Exception\Filter as FilterContract;
 use Viserio\Contracts\Exception\Handler as HandlerContract;
-use Viserio\Contracts\Exception\Transformer as TransformerContract;
 use Viserio\Contracts\Log\Traits\LoggerAwareTrait;
 use Viserio\Exception\Displayers\HtmlDisplayer;
 use Viserio\Exception\Filters\CanDisplayFilter;
 use Viserio\Exception\Filters\VerboseFilter;
-use Viserio\Exception\Transformers\CommandLineTransformer;
 
 class Handler extends ErrorHandler implements HandlerContract
 {
     use ConfigAwareTrait;
-    use ContainerAwareTrait;
     use LoggerAwareTrait;
 
     /**
@@ -65,15 +55,6 @@ class Handler extends ErrorHandler implements HandlerContract
     ];
 
     /**
-     * Exception transformers.
-     *
-     * @var array
-     */
-    protected $transformers = [
-        CommandLineTransformer::class,
-    ];
-
-    /**
      * Exception filters.
      *
      * @var array
@@ -81,19 +62,6 @@ class Handler extends ErrorHandler implements HandlerContract
     protected $filters = [
         VerboseFilter::class,
         CanDisplayFilter::class,
-    ];
-
-    /**
-     * Array of fatal error handlers.
-     *
-     * Override this variable if you want to define more fatal error handlers.
-     *
-     * @return \Symfony\Component\Debug\FatalErrorHandler\FatalErrorHandlerInterface[]
-     */
-    protected $fatalErrorHandlers = [
-        UndefinedFunctionFatalErrorHandler::class,
-        UndefinedMethodFatalErrorHandler::class,
-        ClassNotFoundFatalErrorHandler::class,
     ];
 
     /**
@@ -147,32 +115,6 @@ class Handler extends ErrorHandler implements HandlerContract
     public function getDisplayers(): array
     {
         return $this->displayers;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function addTransformer(TransformerContract $transformer): HandlerContract
-    {
-        $transformerClass = is_object($transformer) ? get_class($transformer) : $transformer;
-
-        if (in_array($transformerClass, $this->transformers)) {
-            $pos = array_search($transformerClass, $this->transformers);
-
-            unset($this->transformers[$pos]);
-        }
-
-        $this->transformers[] = $transformer;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getTransformers(): array
-    {
-        return $this->transformers;
     }
 
     /**
@@ -254,100 +196,6 @@ class Handler extends ErrorHandler implements HandlerContract
     public function unregister()
     {
         restore_error_handler();
-    }
-
-    /**
-     * Convert errors into ErrorException objects.
-     *
-     * This method catches PHP errors and converts them into ErrorException objects;
-     * these ErrorException objects are then thrown and caught by Viserio's
-     * built-in or custom error handlers.
-     *
-     * @param int        $type      The numeric type of the Error
-     * @param string     $message   The error message
-     * @param string     $file      The absolute path to the affected file
-     * @param int        $line      The line number of the error in the affected file
-     * @param null       $context
-     * @param array|null $backtrace
-     *
-     * @throws \ErrorException
-     *
-     * @return bool Returns false when no handling happens so that the PHP engine can handle the error itself
-     *
-     * @internal
-     */
-    public function handleError(
-        int $type,
-        string $message,
-        string $file = '',
-        int $line = 0,
-        $context = null,
-        array $backtrace = null
-    ): void {
-        // Level is the current error reporting level to manage silent error.
-        // Strong errors are not authorized to be silenced.
-        $level = error_reporting() | E_RECOVERABLE_ERROR | E_USER_ERROR | E_DEPRECATED | E_USER_DEPRECATED;
-
-        if ($level) {
-            throw new ErrorException($message, 0, $level, $file, $line);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function handleException($exception)
-    {
-        $exception = $this->prepareAndWrapException($exception);
-
-        $this->report($exception);
-
-        $transformed = $this->getTransformed($exception);
-        $container   = $this->container;
-
-        if (PHP_SAPI === 'cli') {
-            if ($container->has(ConsoleApplication::class)) {
-                $container->get(ConsoleApplication::class)
-                    ->renderException($transformed, new ConsoleOutput());
-            } else {
-                throw $exception;
-            }
-        } else {
-            $response = $this->getPreparedResponse(
-                $container,
-                $exception,
-                $transformed
-            );
-
-            return (string) $response->getBody();
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function handleShutdown()
-    {
-        // If an error has occurred that has not been displayed, we will create a fatal
-        // error exception instance and pass it into the regular exception handling
-        // code so it can be displayed back out to the developer for information.
-        $error = error_get_last();
-
-        if ($error !== null && $this->isFatal($error['type'])) {
-            $this->handleException(
-                // Create a new fatal exception instance from an error array.
-                new FatalThrowableError(
-                    new FatalErrorException(
-                        $error['message'],
-                        $error['type'],
-                        0,
-                        $error['file'],
-                        $error['line'],
-                        0
-                    )
-                )
-            );
-        }
     }
 
     /**
@@ -506,43 +354,6 @@ class Handler extends ErrorHandler implements HandlerContract
     }
 
     /**
-     * Get the transformed exception.
-     *
-     * @param \Throwable|\Exception $exception
-     *
-     * @return \Throwable|\Exception
-     */
-    protected function getTransformed($exception)
-    {
-        $container    = $this->container;
-        $transformers = array_merge(
-            $this->transformers,
-            $container->get(RepositoryContract::class)->get('exception.transformers', [])
-        );
-
-        foreach ($transformers as $transformer) {
-            $transformerClass = is_object($transformer) ? $transformer : $container->get($transformer);
-            $exception        = $transformerClass->transform($exception);
-        }
-
-        return $exception;
-    }
-
-    /**
-     * Determine if the error type is fatal.
-     *
-     * @param int $type
-     *
-     * @return bool
-     *
-     * @codeCoverageIgnore
-     */
-    protected function isFatal(int $type): bool
-    {
-        return in_array($type, [E_ERROR, E_CORE_ERROR, E_COMPILE_ERROR, E_PARSE], true);
-    }
-
-    /**
      * Get a prepared response with the transformed exception.
      *
      * @param \Interop\Container\ContainerInterface $container
@@ -570,83 +381,5 @@ class Handler extends ErrorHandler implements HandlerContract
         }
 
         return $response;
-    }
-
-    /**
-     * Register the PHP error handler.
-     *
-     * @return void
-     */
-    protected function registerErrorHandler(): void
-    {
-        set_error_handler([$this, 'handleError']);
-    }
-
-    /**
-     * Register the PHP exception handler.
-     *
-     * @return void
-     */
-    protected function registerExceptionHandler(): void
-    {
-        if (PHP_SAPI !== 'cli') {
-            ini_set('display_errors', '0');
-        } elseif (! ini_get('log_errors') || ini_get('error_log')) {
-            // CLI - display errors only if they're not already logged to STDERR
-            ini_set('display_errors', '1');
-        }
-
-        set_exception_handler([$this, 'handleException']);
-    }
-
-    /**
-     * Register the PHP shutdown handler.
-     *
-     * @return void
-     */
-    protected function registerShutdownHandler(): void
-    {
-        register_shutdown_function([$this, 'handleShutdown']);
-    }
-
-    /**
-     * Prepare and wrap exception in a fatal error handler.
-     *
-     * @param \Throwable|\Exception $exception
-     *
-     * @return \Symfony\Component\Debug\FatalErrorHandler\FatalErrorHandlerInterface|\Throwable|\Error
-     */
-    protected function prepareAndWrapException($exception)
-    {
-        if (! $exception instanceof Exception) {
-            $exception = new FatalThrowableError($exception);
-        } elseif ($exception instanceof Error) {
-            $exception = new FatalErrorException(
-                $exception->getMessage(),
-                $exception->getCode(),
-                E_ERROR,
-                $exception->getFile(),
-                $exception->getLine(),
-                $exception->getTrace()
-            );
-        }
-
-        if ($exception instanceof FatalErrorException && ! $exception instanceof OutOfMemoryException) {
-            $error = [
-                'type'    => $exception->getSeverity(),
-                'message' => $exception->getMessage(),
-                'file'    => $exception->getFile(),
-                'line'    => $exception->getLine(),
-            ];
-
-            foreach ($this->fatalErrorHandlers as $handler) {
-                if ($e = (new $handler())->handleError($error, $exception)) {
-                    $exception = $e;
-                    break;
-                }
-            }
-        }
-
-        return $exception;
     }
 }
